+++ lib/security.py
@@ -6,11 +6,11 @@ from cryptography.hazmat.primitives.kdf.pbkdf2 import PBKDF2HMAC
 from .consts import CONFIG_PARAM_SECRET_CONST
 
 
-def is_password_encrypted(password):
+def is_password_encrypted(password: str) -> bool:
     return password is not None and password[-4:] == '????'
 
 
-def set_up_encryption(server_name, port):
+def set_up_encryption(server_name: str, port: int) -> Fernet:
     salt = bytes(port)
     # TODO: rewrite to AES
     kdf = PBKDF2HMAC(algorithm=hashes.SHA256(),
@@ -22,13 +22,13 @@ def set_up_encryption(server_name, port):
     return f
 
 
-def decrypt_password(password, server_name, port):
+def decrypt_password(password: str, server_name: str, port: int) -> str:
     f = set_up_encryption(server_name, port)
     password = f.decrypt(password.encode('UTF-8'))
     return password.decode('UTF-8')
 
 
-def encrypt_password(password, server_name, port):
+def encrypt_password(password: str, server_name: str, port: int) -> str:
     f = set_up_encryption(server_name, port)
     password = f.encrypt(password.encode('UTF-8'))
     # TODO fix this hack for secret strings