+++ main.py
@@ -1,15 +1,17 @@
 import argparse
 import datetime
 import json
-import psutil
 import os
 import sys
 import time
 
-
 import pika
+import psutil
 from telegram import InlineKeyboardButton, InlineKeyboardMarkup
 from telegram.ext import CommandHandler, Filters, MessageHandler, Updater, CallbackQueryHandler
+from telegram.ext.callbackcontext import CallbackContext
+from telegram.update import Update
+from typing import List, Dict
 
 from lib.config import Config
 from lib.consts import *
@@ -25,6 +27,7 @@ global creation_process
 global deletion_process
 global feedback_process
 global feedback_reading
+global feedback_replying
 global characters
 global updater
 global queue_logger
@@ -37,7 +40,7 @@ global startup_time
 global user_settings
 
 
-def get_locale(update, chat_id=None):
+def get_locale(update: Update, chat_id: int = None):
     global user_locales
     global translations
     if update is not None or chat_id is not None:
@@ -51,12 +54,12 @@ def get_locale(update, chat_id=None):
                 locale = update.message.from_user.language_code
             user_locales[chat_id] = locale
 
-        if locale in translations.keys():
+        if locale in translations:
             return translations[locale]
     return translations['ru']
 
 
-def start(update, context):
+def start(update: Update, context: CallbackContext):
     global telegram_logger
     trans = get_locale(update)
     msg = trans.get_message(M_ABOUT_ME)
@@ -65,7 +68,7 @@ def start(update, context):
     telegram_logger.info("Proceed start command from user {0}".format(update.effective_chat.id))
 
 
-def pretty_menu(menu):
+def pretty_menu(menu: List):
     res = [[]]
     cur_menu_len = 0
     for i in menu:
@@ -77,16 +80,16 @@ def pretty_menu(menu):
     return res
 
 
-def class_keyboard(trans):
+def class_keyboard(trans: L18n):
     keyboard = []
-    if class_list is None:
+    if not class_list:
         return None
     for i in class_list:
         keyboard.append(InlineKeyboardButton(trans.get_message(i), callback_data="class_" + str(i)))
     return pretty_menu(keyboard)
 
 
-def main_keyboard(chat_id, trans):
+def main_keyboard(chat_id: int, trans: L18n):
     global config
     keyboard = [
         InlineKeyboardButton(trans.get_message(M_NEW_CHARACTER), callback_data=MAIN_MENU_CREATE),
@@ -102,7 +105,7 @@ def main_keyboard(chat_id, trans):
     return pretty_menu(keyboard)
 
 
-def admin_keyboard(trans):
+def admin_keyboard(trans: L18n):
     keyboard = [
         InlineKeyboardButton(trans.get_message(M_SERVER_STATS), callback_data=ADMIN_MENU_STATS),
         InlineKeyboardButton(trans.get_message(M_BOT_STATS), callback_data=ADMIN_MENU_BOT_STATS),
@@ -113,15 +116,16 @@ def admin_keyboard(trans):
     return pretty_menu(keyboard)
 
 
-def read_keyboard(trans):
+def read_keyboard(trans: L18n):
     keyboard = [
-        InlineKeyboardButton(trans.get_message(M_FEEDBACK_DONE), callback_data=READ_MENU_DONE)
+        InlineKeyboardButton(trans.get_message(M_FEEDBACK_DONE), callback_data=READ_MENU_DONE),
+        InlineKeyboardButton(trans.get_message(M_FEEDBACK_REPLY), callback_data=READ_MENU_REPLY)
     ]
 
     return pretty_menu(keyboard)
 
 
-def shutdown_keyboard(trans):
+def shutdown_keyboard(trans: L18n):
     keyboard = [
          InlineKeyboardButton(trans.get_message(M_SHUTDOWN_NORMAL), callback_data=SHUTDOWN_MENU_NORMAL),
          InlineKeyboardButton(trans.get_message(M_SHUTDOWN_IMMEDIATE), callback_data=SHUTDOWN_MENU_IMMEDIATE),
@@ -131,16 +135,18 @@ def shutdown_keyboard(trans):
     return pretty_menu(keyboard)
 
 
-def locale_keyboard():
+def locale_keyboard(trans: L18n):
     global translations
     keyboard = []
     for i in translations:
-        keyboard.append(InlineKeyboardButton(i, callback_data="LOCALE_" + i))
+        keyboard.append(InlineKeyboardButton(i, callback_data=LOCALE_PREFIX + i))
+    keyboard.append(InlineKeyboardButton(trans.get_message(M_DYNAMIC_LOCALE),
+                                         callback_data=LOCALE_PREFIX + M_DYNAMIC_LOCALE))
 
     return pretty_menu(keyboard)
 
 
-def status(update, context):
+def status(update: Update, context: CallbackContext):
     global telegram_logger
     trans = get_locale(update)
     cmd = {"user_id": update.effective_chat.id, "cmd_type": CMD_GET_CHARACTER_STATUS, "locale": trans.code}
@@ -150,7 +156,35 @@ def status(update, context):
     telegram_logger.info("Proceed status command from user {0}".format(update.effective_chat.id))
 
 
-def create(update, context):
+def reset_process(user_id: int,
+                  creation: bool = False,
+                  deletion: bool = False,
+                  feedback_send: bool = False,
+                  feedback_read: bool = False,
+                  feedback_reply: bool = False):
+    global creation_process
+    global deletion_process
+    global feedback_process
+    global feedback_reading
+    global feedback_replying
+    if not creation:
+        if user_id in creation_process:
+            del creation_process[user_id]
+    if not deletion:
+        if user_id in deletion_process:
+            del deletion_process[user_id]
+    if not feedback_send:
+        if user_id in feedback_process:
+            del feedback_process[user_id]
+    if not feedback_read:
+        if user_id in feedback_reading:
+            del feedback_reading[user_id]
+    if not feedback_replying:
+        if user_id in feedback_replying:
+            del feedback_replying[user_id]
+
+
+def create(update: Update, context: CallbackContext):
     global creation_process
     global telegram_logger
     trans = get_locale(update)
@@ -158,6 +192,7 @@ def create(update, context):
     if keyboard is not None:
         msg = trans.get_message(M_CHOOSE_CLASS)
         creation_process[update.effective_chat.id] = {"stage": STAGE_SELECT_CLASS}
+        reset_process(user_id=update.effective_chat.id, creation=True)
         telegram_logger.info("Initialized character creation from user {0}".format(update.effective_chat.id))
     else:
         msg = trans.get_message(M_REPEAT_LATER)
@@ -168,63 +203,75 @@ def create(update, context):
     context.bot.send_message(chat_id=update.effective_chat.id, text=msg, reply_markup=reply_markup)
 
 
-def delete(update, context):
+def delete(update: Update, context: CallbackContext):
     global deletion_process
     global telegram_logger
     deletion_process[update.effective_chat.id] = {"stage": STAGE_CONFIRM_DELETION}
+    reset_process(user_id=update.effective_chat.id, deletion=True)
     trans = get_locale(update)
     msg = trans.get_message(M_PRINT_CONFIRM)
     context.bot.send_message(chat_id=update.effective_chat.id, text=msg)
     telegram_logger.info("Initialized character deletion from user {0}".format(update.effective_chat.id))
 
 
-def settings(update, context):
+def settings(update: Update, context: CallbackContext):
     global telegram_logger
     global translations
     trans = get_locale(update)
     msg = trans.get_message(M_CHOOSE_LANGUAGE)
-    keyboard = locale_keyboard()
+    keyboard = locale_keyboard(trans)
     reply_markup = InlineKeyboardMarkup(keyboard)
     context.bot.send_message(chat_id=update.effective_chat.id, text=msg, reply_markup=reply_markup)
     telegram_logger.info("Sent language settings menu to user {0}".format(update.effective_chat.id))
 
 
-def feedback(update, context):
+def feedback(update: Update, context: CallbackContext):
     global telegram_logger
     global translations
     global feedback_process
     trans = get_locale(update)
     msg = trans.get_message(M_FEEDBACK_PROMPT)
     feedback_process[update.effective_chat.id] = 1
+    reset_process(user_id=update.effective_chat.id, feedback_send=True)
     context.bot.send_message(chat_id=update.effective_chat.id, text=msg)
     telegram_logger.info("Sent feedback prompt to user {0}".format(update.effective_chat.id))
 
 
-def set_locale(update, context):
+def set_locale(update: Update, context: CallbackContext):
     global telegram_logger
     global translations
     global user_locales
     global user_settings
     language = update["callback_query"]["data"][7:]
-    if language in translations.keys():
+    if language in translations:
         user_locales[update.effective_chat.id] = language
-        user_settings.set(update.effective_chat.id, language)
+        user_settings.set_locale(update.effective_chat.id, language)
         trans = get_locale(update)
         msg = trans.get_message(M_LANGUAGE_CHOSEN).format(language)
         keyboard = main_keyboard(update.effective_chat.id, trans)
         reply_markup = InlineKeyboardMarkup(keyboard)
         context.bot.send_message(chat_id=update.effective_chat.id, text=msg, reply_markup=reply_markup)
         telegram_logger.info("Set locale {1} for user {0}".format(update.effective_chat.id, language))
+    elif language == M_DYNAMIC_LOCALE:
+        del user_locales[update.effective_chat.id]
+        user_settings.delete_locale(update.effective_chat.id)
+        trans = get_locale(update)
+        msg = trans.get_message(M_LANGUAGE_RESET)
+        keyboard = main_keyboard(update.effective_chat.id, trans)
+        reply_markup = InlineKeyboardMarkup(keyboard)
+        context.bot.send_message(chat_id=update.effective_chat.id, text=msg, reply_markup=reply_markup)
+        telegram_logger.info("Deleted locale settings for user {0}".format(update.effective_chat.id))
     else:
         trans = get_locale(update)
         msg = trans.get_message(M_INCORRECT_LANGUAGE).format(language)
         keyboard = main_keyboard(update.effective_chat.id, trans)
         reply_markup = InlineKeyboardMarkup(keyboard)
         context.bot.send_message(chat_id=update.effective_chat.id, text=msg, reply_markup=reply_markup)
-        telegram_logger.error("Can't set not existent locale {1} for user {0}".format(update.effective_chat.id, language))
+        telegram_logger.error("Can't set not existent locale {1} for user {0}".format(update.effective_chat.id,
+                                                                                      language))
 
 
-def about(update, context):
+def about(update: Update, context: CallbackContext):
     global telegram_logger
     trans = get_locale(update)
     msg = trans.get_message(M_ABOUT_TEXT)
@@ -234,7 +281,7 @@ def about(update, context):
     telegram_logger.info("Sent \"About\" to user {0}".format(update.effective_chat.id))
 
 
-def admin(update, context):
+def admin(update: Update, context: CallbackContext):
     global telegram_logger
     global config
     if update.effective_chat.id in config.admin_list:
@@ -248,7 +295,7 @@ def admin(update, context):
         telegram_logger.error("Illegal access to \"Admin menu\" from user {0}".format(update.effective_chat.id))
 
 
-def shutdown(update, context):
+def shutdown(update: Update, context: CallbackContext):
     global telegram_logger
     global config
     if update.effective_chat.id in config.admin_list:
@@ -262,17 +309,15 @@ def shutdown(update, context):
         telegram_logger.error("Illegal access to \"shutdown menu\" from user {0}".format(update.effective_chat.id))
 
 
-def get_feedback(update, context):
+def get_feedback(update: Update, context: CallbackContext):
     global telegram_logger
     global config
     if update.effective_chat.id in config.admin_list:
         trans = get_locale(update)
-        msg = trans.get_message(M_GET_FEEDBACK)
+        msg = trans.get_message(M_CMD_SENT)
         cmd = {"user_id": update.effective_chat.id, "cmd_type": CMD_GET_FEEDBACK, "locale": trans.code}
         enqueue_command(cmd, True)
-        keyboard = admin_keyboard(trans)
-        reply_markup = InlineKeyboardMarkup(keyboard)
-        context.bot.send_message(chat_id=update.effective_chat.id, text=msg, reply_markup=reply_markup)
+        context.bot.send_message(chat_id=update.effective_chat.id, text=msg)
         telegram_logger.info("Sent get_feedback from user {0}".format(update.effective_chat.id))
     else:
         telegram_logger.error("Illegal access to \"feedback menu\" from user {0}".format(update.effective_chat.id))
@@ -292,7 +337,7 @@ def ask_server_stats(update, context):
         telegram_logger.error("Illegal access to \"ask_server_stats\" from user {0}".format(update.effective_chat.id))
 
 
-def show_bot_stats(update, context):
+def show_bot_stats(update: Update, context: CallbackContext):
     global telegram_logger
     global config
     global startup_time
@@ -320,7 +365,7 @@ def show_bot_stats(update, context):
         telegram_logger.error("Illegal access to \"show_bot_stats\" from user {0}".format(update.effective_chat.id))
 
 
-def send_shutdown_immediate(update, context):
+def send_shutdown_immediate(update: Update, context: CallbackContext):
     global telegram_logger
     global config
     if update.effective_chat.id in config.admin_list:
@@ -335,7 +380,7 @@ def send_shutdown_immediate(update, context):
                               format(update.effective_chat.id))
 
 
-def send_shutdown_bot(update, context):
+def send_shutdown_bot(update: Update, context: CallbackContext):
     global telegram_logger
     global config
     global is_shutdown
@@ -349,7 +394,7 @@ def send_shutdown_bot(update, context):
         telegram_logger.error("Illegal access to \"send_shutdown_bot\" from user {0}".format(update.effective_chat.id))
 
 
-def send_shutdown_normal(update, context):
+def send_shutdown_normal(update: Update, context: CallbackContext):
     global telegram_logger
     global config
     if update.effective_chat.id in config.admin_list:
@@ -364,13 +409,13 @@ def send_shutdown_normal(update, context):
                               format(update.effective_chat.id))
 
 
-def class_menu(update, context):
+def class_menu(update: Update, context: CallbackContext):
     global creation_process
     global telegram_logger
     global class_descriptions
     is_correct = False
     is_restart = False
-    if update.effective_chat.id in creation_process.keys():
+    if update.effective_chat.id in creation_process:
         if creation_process[update.effective_chat.id]["stage"] == STAGE_SELECT_CLASS:
             is_correct = True
         elif creation_process[update.effective_chat.id]["stage"] == STAGE_CHOOSE_NAME:
@@ -379,6 +424,7 @@ def class_menu(update, context):
         char_class = update["callback_query"]["data"][6:]
         creation_process[update.effective_chat.id]["class"] = char_class
         creation_process[update.effective_chat.id]["stage"] = STAGE_CHOOSE_NAME
+        reset_process(user_id=update.effective_chat.id, creation=True)
         trans = get_locale(update)
         descr_code = char_class + "_description"
         if trans.is_message_exists(descr_code):
@@ -399,11 +445,11 @@ def class_menu(update, context):
         create(update, context)
 
 
-def read_done(update, context):
+def read_done(update: Update, context: CallbackContext):
     global feedback_reading
     global telegram_logger
     is_correct = False
-    if update.effective_chat.id in feedback_reading.keys():
+    if update.effective_chat.id in feedback_reading:
         is_correct = True
     if is_correct:
         trans = get_locale(update)
@@ -416,12 +462,32 @@ def read_done(update, context):
                              format(feedback_reading[update.effective_chat.id], update.effective_chat.id))
         del feedback_reading[update.effective_chat.id]
     else:
-        telegram_logger.warning("Reading done failed".
+        telegram_logger.warning("Message expire (on read done), requested new message".
                                 format(update.effective_chat.id))
-        start(update, context)
+        get_feedback(update, context)
 
 
-def main_menu(update, context):
+def read_reply(update: Update, context: CallbackContext):
+    global feedback_reading
+    global feedback_replying
+    global telegram_logger
+    is_correct = False
+    if update.effective_chat.id in feedback_reading:
+        is_correct = True
+    if is_correct:
+        trans = get_locale(update)
+        msg = trans.get_message(M_PRINT_REPLY)
+        context.bot.send_message(chat_id=update.effective_chat.id, text=msg)
+        telegram_logger.info("Ask for reply on message id {0} user {1}".
+                             format(feedback_reading[update.effective_chat.id], update.effective_chat.id))
+        feedback_replying[update.effective_chat.id] = 1
+    else:
+        telegram_logger.warning("Message expire (on read reply), requested new message".
+                                format(update.effective_chat.id))
+        get_feedback(update, context)
+
+
+def main_menu(update: Update, context: CallbackContext):
     global telegram_logger
     global config
     cur_item = update["callback_query"]["data"]
@@ -453,7 +519,7 @@ def main_menu(update, context):
         context.bot.send_message(chat_id=update.effective_chat.id, text="Unknown command")
 
 
-def admin_menu(update, context):
+def admin_menu(update: Update, context: CallbackContext):
     global telegram_logger
     cur_item = update["callback_query"]["data"]
     telegram_logger.info("Received command {0} from user {1} in admin menu".
@@ -472,7 +538,7 @@ def admin_menu(update, context):
         context.bot.send_message(chat_id=update.effective_chat.id, text="Unknown command")
 
 
-def shutdown_menu(update, context):
+def shutdown_menu(update: Update, context: CallbackContext):
     global telegram_logger
     cur_item = update["callback_query"]["data"]
     telegram_logger.info("Received command {0} from user {1} in shutdown menu".
@@ -489,20 +555,22 @@ def shutdown_menu(update, context):
         context.bot.send_message(chat_id=update.effective_chat.id, text="Unknown command")
 
 
-def read_menu(update, context):
+def read_menu(update: Update, context: CallbackContext):
     global telegram_logger
     cur_item = update["callback_query"]["data"]
     telegram_logger.info("Received command {0} from user {1} in read menu".
                          format(cur_item, update.effective_chat.id))
     if cur_item == READ_MENU_DONE:
         read_done(update, context)
+    elif cur_item == READ_MENU_REPLY:
+        read_reply(update, context)
     else:
         telegram_logger.error("Received unknown command {0} from user {1} in read menu".
                               format(cur_item, update.effective_chat.id))
         context.bot.send_message(chat_id=update.effective_chat.id, text="Unknown command")
 
 
-def enqueue_command(obj, system=False):
+def enqueue_command(obj: Dict, system: bool = False):
     global queue_logger
     global config
     if system:
@@ -525,15 +593,17 @@ def enqueue_command(obj, system=False):
         queue_logger.critical("Error {2} when Sent command {0} in queue {1}".format(msg_body, queue_name, exc))
 
 
-def echo(update, context):
+def echo(update: Update, context: CallbackContext):
     global creation_process
     global deletion_process
     global feedback_process
+    global feedback_reading
+    global feedback_replying
     global telegram_logger
     telegram_logger.debug("Echo: update: {0}, context {1}".format(update, context))
     trans = get_locale(update, update.effective_chat.id)
     is_correct = False
-    if update.effective_chat.id in creation_process.keys():
+    if update.effective_chat.id in creation_process:
         if creation_process[update.effective_chat.id]["stage"] == STAGE_CHOOSE_NAME:
             if CHARACTER_NAME_MAX_LENGTH < len(update["message"]["text"]):
                 msg = trans.get_message(M_NAME_TOO_LONG).format(CHARACTER_NAME_MAX_LENGTH)
@@ -550,7 +620,7 @@ def echo(update, context):
                "class": creation_process[update.effective_chat.id].get("class"),
                "locale": trans.code}
         enqueue_command(cmd)
-    elif update.effective_chat.id in deletion_process.keys():
+    elif update.effective_chat.id in deletion_process:
         if deletion_process[update.effective_chat.id]["stage"] == STAGE_CONFIRM_DELETION:
             if update["message"]["text"] == "CONFIRM":
                 cmd = {"user_id": update.effective_chat.id, "cmd_type": CMD_DELETE_CHARACTER, "locale": trans.code}
@@ -562,7 +632,7 @@ def echo(update, context):
                 msg = trans.get_message(M_CANCEL_REQUEST)
                 context.bot.send_message(chat_id=update.effective_chat.id, text=msg)
                 start(update, context)
-    elif update.effective_chat.id in feedback_process.keys():
+    elif update.effective_chat.id in feedback_process:
         if len(update["message"]["text"]) <= MAX_FEEDBACK_LENGTH:
             del feedback_process[update.effective_chat.id]
             cmd = {"user_id": update.effective_chat.id, "cmd_type": CMD_FEEDBACK, "locale": trans.code,
@@ -574,12 +644,21 @@ def echo(update, context):
             msg = trans.get_message(M_FEEDBACK_TOO_LONG).format(MAX_FEEDBACK_LENGTH)
             reply_markup = InlineKeyboardMarkup(main_keyboard(None, trans))
             context.bot.send_message(chat_id=update.effective_chat.id, text=msg, reply_markup=reply_markup)
+    elif update.effective_chat.id in feedback_replying:
+        if update.effective_chat.id in feedback_reading:
+            cmd = {"user_id": update.effective_chat.id, "cmd_type": CMD_REPLY_FEEDBACK, "locale": trans.code,
+                   "message": update["message"]["text"], "message_id": feedback_reading.get(update.effective_chat.id)}
+            msg = trans.get_message(M_FEEDBACK_SUCCESS)
+            context.bot.send_message(chat_id=update.effective_chat.id, text=msg)
+            enqueue_command(cmd)
+        else:
+            get_feedback(update, context)
 
     else:
         telegram_logger.info("User {0} sent message {1}".format(update.effective_chat.id, update.message.text))
 
 
-def class_list_callback(ch, method, properties, body):
+def class_list_callback(ch, method: pika.spec.Basic.Deliver, properties: pika.BasicProperties, body: bytes):
     global class_list
     global translations
     buf = json.loads(body).get("class_list")
@@ -587,22 +666,22 @@ def class_list_callback(ch, method, properties, body):
     for i in buf:
         class_list.append(i)
         for j in buf[i]:
-            if j in translations.keys():
+            if j in translations:
                 translations[j].add_message(i, buf[i][j])
 
 
-def class_description_callback(ch, method, properties, body):
+def class_description_callback(ch, method: pika.spec.Basic.Deliver, properties: pika.BasicProperties, body: bytes):
     global class_descriptions
     global translations
     class_name = json.loads(body).get("class_name")
     class_description = json.loads(body).get("class_description")
     class_stats = json.loads(body).get("class_stats")
     locale = json.loads(body).get("locale")
-    if locale in translations.keys():
+    if locale in translations:
         translations[locale].add_message(str(class_name) + "_description", class_description + chr(10) + class_stats)
 
 
-def dict_response_callback(ch, method, properties, body):
+def dict_response_callback(ch, method: pika.spec.Basic.Deliver, properties: pika.BasicProperties, body: bytes):
     global queue_logger
     global feedback_reading
     queue_logger.info("Received server command " + str(body) + ", started callback")
@@ -623,6 +702,7 @@ def dict_response_callback(ch, method, properties, body):
     elif cmd_type == CMD_SENT_FEEDBACK:
         reply_markup = InlineKeyboardMarkup(read_keyboard(trans))
         feedback_reading[chat_id] = msg.get("message_id")
+        reset_process(user_id=chat_id, feedback_read=True)
         updater.dispatcher.bot.send_message(chat_id=chat_id,
                                             text=trans.get_message(M_FEEDBACK_STRING).
                                             format(msg.get("user_sent_id"), msg.get("user_sent_nick"),
@@ -634,7 +714,7 @@ def dict_response_callback(ch, method, properties, body):
         queue_logger.error("Received unknown server command " + str(body) + ", started callback")
 
 
-def cmd_response_callback(ch, method, properties, body):
+def cmd_response_callback(ch, method: pika.spec.Basic.Deliver, properties: pika.BasicProperties, body: bytes):
     global creation_process
     global deletion_process
     global updater
@@ -650,17 +730,18 @@ def cmd_response_callback(ch, method, properties, body):
         elif msg.get("cmd_type") == CMD_FEEDBACK_RECEIVE:
             updater.dispatcher.bot.send_message(chat_id=chat_id, text=trans.get_message(M_FEEDBACK_SUCCESS),
                                                 reply_markup=reply_markup)
+        elif msg.get("cmd_type") == CMD_REPLY_FEEDBACK:
+            updater.dispatcher.bot.send_message(chat_id=chat_id,
+                                                text=trans.get_message(M_ADMIN_ANSWER).format(msg.get("message")),
+                                                reply_markup=reply_markup)
         else:
             updater.dispatcher.bot.send_message(chat_id=chat_id, text=msg.get("message"), reply_markup=reply_markup)
             queue_logger.info("Sent message {0}, received from server to user {1}".format(msg.get("message"), chat_id))
         # clear current operations state, if any
-        if chat_id in creation_process.keys():
-            del creation_process[chat_id]
-        if chat_id in deletion_process.keys():
-            del deletion_process[chat_id]
+        reset_process(user_id=chat_id)
 
 
-def get_mq_connect(mq_config):
+def get_mq_connect(mq_config: Config):
     if mq_config.queue_password is None:
         return pika.BlockingConnection(pika.ConnectionParameters(host=mq_config.queue_host, port=mq_config.queue_port))
     else:
@@ -676,6 +757,7 @@ def main():
     global deletion_process
     global feedback_process
     global feedback_reading
+    global feedback_replying
     global characters
     global out_channel
     global updater
@@ -689,12 +771,13 @@ def main():
     global user_settings
 
     is_shutdown = False
-    class_list = None
+    class_list = []
     class_descriptions = {}
     creation_process = {}
     deletion_process = {}
     feedback_process = {}
     feedback_reading = {}
+    feedback_replying = {}
     characters = {}
     user_locales = {}
     translations = {}
@@ -704,7 +787,6 @@ def main():
     parser.add_argument("--config", '-cfg', help="Path to config file", action="store", default="cfg//main.json")
     parser.add_argument("--test_users", help="Number of test users of each class created", action="store", default=None)
     parser.add_argument("--delay", help="Number of test users of each class created", action="store", default=None)
-    parser.add_argument("--db", help="Path to the user settings storage", action="store", default="user_settings.db")
     args = parser.parse_args()
     if args.delay is not None:
         time.sleep(int(args.delay))
@@ -716,7 +798,8 @@ def main():
     # set_basic_logging(config.log_level)
 
     user_settings = Persist(config)
-    user_locales = user_settings.get_all()
+    user_settings.check_version()
+    user_locales = user_settings.get_all_locale()
 
     for dirpath, dirnames, filenames in os.walk("l18n"):
         for lang_file in filenames:
@@ -735,7 +818,7 @@ def main():
     main_menu_handler = CallbackQueryHandler(main_menu, pattern="main_")
     admin_menu_handler = CallbackQueryHandler(admin_menu, pattern="admin_")
     shutdown_menu_handler = CallbackQueryHandler(shutdown_menu, pattern="shutdown_")
-    locale_menu_handler = CallbackQueryHandler(set_locale, pattern="LOCALE_")
+    locale_menu_handler = CallbackQueryHandler(set_locale, pattern=LOCALE_PREFIX)
     read_menu_handler = CallbackQueryHandler(read_menu, pattern="confirm_")
     echo_handler = MessageHandler(Filters.text & (~Filters.command), echo)
     dispatcher.add_handler(start_handler)
@@ -767,7 +850,7 @@ def main():
         out_channel.basic_consume(queue=QUEUE_NAME_DICT, on_message_callback=dict_response_callback, auto_ack=True)
 
         for method_frame, properties, body in out_channel.consume(QUEUE_NAME_DICT, inactivity_timeout=1):
-            if class_list is not None:
+            if class_list:
                 break
         out_channel.cancel()
         logger.info("Class list received")
@@ -796,7 +879,7 @@ def main():
                                                                                          method_frame.delivery_tag))
                     cmd_response_callback(None, method_frame, properties, body)
                     out_channel.basic_ack(method_frame.delivery_tag)
-                    logger.info("Received user message " + str(body) + " with delivery_tag " +
+                    logger.info("User message " + str(body) + " with delivery_tag " +
                                 str(method_frame.delivery_tag) + " acknowledged")
                 else:
                     logger.info("No more messages in {0}".format(QUEUE_NAME_RESPONSES))